LayoutTests/fast/text/lang-mapped-to-webkit-locale-expected.txt

This is a test for bug 67586. HTML attributes lang and xml:lang are mapped to '-webkit-locale' to use their value in font fallback.

PASS getLangOfNode('x1') is "ja"
PASS getLangOfNode('x2') is "ja"
PASS getLangOfNode('x3') is "ja"
PASS getLangOfNode('l1') is "zh"
PASS getLangOfNode('l2') is "zh"
PASS getLangOfNode('l3') is "zh"
PASS getLangOfNode('m1') is "ko"
PASS getLangOfNode('m2') is "fr"
PASS getLangOfNode('m3') is "fr"
PASS getLangOfNode('n1') is "fr"
PASS getLangOfNode('n2') is "fr"
PASS getLangOfNode('n3') is "sv"
PASS getLangOfNode('p1') is "ja"
PASS getLangOfNode('p2') is "auto"
PASS getLangOfNode('q1') is "ja"
PASS getLangOfNode('q2') is "auto"
PASS getLangOfNode('q3') is "auto"
PASS getLangOfNode('q4') is "ar"
PASS getLangOfNode('q5') is "auto"
PASS getLangOfNode('q6') is "'  '"
PASS getLangOfNode('q7') is "auto"
PASS getLangOfNode('q8') is "xyzzy"
PASS getLangOfNode('q9') is "'][;][['"
PASS successfullyParsed is true

TEST COMPLETE
